+++ src/test/java/calculation/HalfDownDefeatedCalculationShould.java
@@ -17,12 +17,12 @@ public class HalfDownDefeatedCalculationShould {
   @Parameterized.Parameters
   public static Collection<Object[]> data() {
     return Arrays.asList(new Object[][] {
-            { 10.0f, 0.0 }, { 20.0f, 5.0 }, { 30.0f, 15 }, { 40.0f, 25 }, { 50.0f, 35 }
+            { 10.0, 0.0 }, { 20.0, 5.0 }, { 30.0, 15 }, { 40.0, 25 }, { 50.0, 35 }
     });
   }
 
   @Parameterized.Parameter
-  public /* NOT private */ float defeatedPercentage;
+  public /* NOT private */ double defeatedPercentage;
 
   @Parameterized.Parameter(1)
   public /* NOT private */ double expectedSurvivorsPercentage;
